// src/generated/t_tech/invest/grpc/operations.ts
export interface BrokerReportResponse {
  generateBrokerReportResponse?: GenerateBrokerReportResponse | undefined;
  getBrokerReportResponse?: GetBrokerReportResponse | undefined;
}

export interface GenerateBrokerReportRequest {
  /** Идентификатор счета клиента. */
  accountId: string;
  /** Начало периода по UTC. */
  from:
    | Date
    | undefined;
  /** Окончание периода по UTC. */
  to: Date | undefined;
}

export interface GenerateBrokerReportResponse {
  /** Идентификатор задачи формирования брокерского отчета. */
  taskId: string;
}

export interface GetBrokerReportRequest {
  /** Идентификатор задачи формирования брокерского отчета. */
  taskId: string;
  /** Номер страницы отчета, начинается с 1. Значение по умолчанию — 0. */
  page?: number | undefined;
}

export interface GetBrokerReportResponse {
  brokerReport: BrokerReport[];
  /** Количество записей в отчете. */
  itemsCount: number;
  /** Количество страниц с данными отчета, начинается с 0. */
  pagesCount: number;
  /** Текущая страница, начинается с 0. */
  page: number;
  /** Идентификатор задачи формирования брокерского отчета. */
  taskId: string;
}

export interface BrokerReport {
  /** Номер сделки. */
  tradeId: string;
  /** Номер поручения. */
  orderId: string;
  /** FIGI-идентификаторинструмента. */
  figi: string;
  /** Признак исполнения. */
  executeSign: string;
  /** Дата и время заключения по UTC. */
  tradeDatetime:
    | Date
    | undefined;
  /** Торговая площадка. */
  exchange: string;
  /** Режим торгов. */
  classCode: string;
  /** Вид сделки. */
  direction: string;
  /** Сокращенное наименование актива. */
  name: string;
  /** Код актива. */
  ticker: string;
  /** Цена за единицу. */
  price:
    | MoneyValue
    | undefined;
  /** Количество. */
  quantity: number;
  /** Сумма без НКД. */
  orderAmount:
    | MoneyValue
    | undefined;
  /** НКД. */
  aciValue:
    | Quotation
    | undefined;
  /** Сумма сделки. */
  totalOrderAmount:
    | MoneyValue
    | undefined;
  /** Комиссия брокера. */
  brokerCommission:
    | MoneyValue
    | undefined;
  /** Комиссия биржи. */
  exchangeCommission:
    | MoneyValue
    | undefined;
  /** Комиссия клирингового центра. */
  exchangeClearingCommission:
    | MoneyValue
    | undefined;
  /** Ставка РЕПО, %. */
  repoRate:
    | Quotation
    | undefined;
  /** Контрагент или брокерарокер. */
  party: string;
  /** Дата расчетов по UTC. */
  clearValueDate:
    | Date
    | undefined;
  /** Дата поставки по UTC. */
  secValueDate:
    | Date
    | undefined;
  /** Статус брокера. */
  brokerStatus: string;
  /** Тип договора. */
  separateAgreementType: string;
  /** Номер договора. */
  separateAgreementNumber: string;
  /** Дата договора. */
  separateAgreementDate: string;
  /** Тип расчета по сделке. */
  deliveryType: string;
}
// ...
export const BrokerReportResponse: MessageFns<BrokerReportResponse> = {
  encode(message: BrokerReportResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.generateBrokerReportResponse !== undefined) {
      GenerateBrokerReportResponse.encode(message.generateBrokerReportResponse, writer.uint32(10).fork()).join();
    }
    if (message.getBrokerReportResponse !== undefined) {
      GetBrokerReportResponse.encode(message.getBrokerReportResponse, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BrokerReportResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBrokerReportResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.generateBrokerReportResponse = GenerateBrokerReportResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.getBrokerReportResponse = GetBrokerReportResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BrokerReportResponse {
    return {
      generateBrokerReportResponse: isSet(object.generateBrokerReportResponse)
        ? GenerateBrokerReportResponse.fromJSON(object.generateBrokerReportResponse)
        : isSet(object.generate_broker_report_response)
        ? GenerateBrokerReportResponse.fromJSON(object.generate_broker_report_response)
        : undefined,
      getBrokerReportResponse: isSet(object.getBrokerReportResponse)
        ? GetBrokerReportResponse.fromJSON(object.getBrokerReportResponse)
        : isSet(object.get_broker_report_response)
        ? GetBrokerReportResponse.fromJSON(object.get_broker_report_response)
        : undefined,
    };
  },

  toJSON(message: BrokerReportResponse): unknown {
    const obj: any = {};
    if (message.generateBrokerReportResponse !== undefined) {
      obj.generateBrokerReportResponse = GenerateBrokerReportResponse.toJSON(message.generateBrokerReportResponse);
    }
    if (message.getBrokerReportResponse !== undefined) {
      obj.getBrokerReportResponse = GetBrokerReportResponse.toJSON(message.getBrokerReportResponse);
    }
    return obj;
  },

  create(base?: DeepPartial<BrokerReportResponse>): BrokerReportResponse {
    return BrokerReportResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<BrokerReportResponse>): BrokerReportResponse {
    const message = createBaseBrokerReportResponse();
    message.generateBrokerReportResponse =
      (object.generateBrokerReportResponse !== undefined && object.generateBrokerReportResponse !== null)
        ? GenerateBrokerReportResponse.fromPartial(object.generateBrokerReportResponse)
        : undefined;
    message.getBrokerReportResponse =
      (object.getBrokerReportResponse !== undefined && object.getBrokerReportResponse !== null)
        ? GetBrokerReportResponse.fromPartial(object.getBrokerReportResponse)
        : undefined;
    return message;
  },
};
// ...
export const GenerateBrokerReportResponse: MessageFns<GenerateBrokerReportResponse> = {
  encode(message: GenerateBrokerReportResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.taskId !== "") {
      writer.uint32(10).string(message.taskId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GenerateBrokerReportResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGenerateBrokerReportResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.taskId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GenerateBrokerReportResponse {
    return {
      taskId: isSet(object.taskId)
        ? globalThis.String(object.taskId)
        : isSet(object.task_id)
        ? globalThis.String(object.task_id)
        : "",
    };
  },

  toJSON(message: GenerateBrokerReportResponse): unknown {
    const obj: any = {};
    if (message.taskId !== "") {
      obj.taskId = message.taskId;
    }
    return obj;
  },

  create(base?: DeepPartial<GenerateBrokerReportResponse>): GenerateBrokerReportResponse {
    return GenerateBrokerReportResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GenerateBrokerReportResponse>): GenerateBrokerReportResponse {
    const message = createBaseGenerateBrokerReportResponse();
    message.taskId = object.taskId ?? "";
    return message;
  },
};
// ...
export const GetBrokerReportResponse: MessageFns<GetBrokerReportResponse> = {
  encode(message: GetBrokerReportResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.brokerReport) {
      BrokerReport.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.itemsCount !== 0) {
      writer.uint32(16).int32(message.itemsCount);
    }
    if (message.pagesCount !== 0) {
      writer.uint32(24).int32(message.pagesCount);
    }
    if (message.page !== 0) {
      writer.uint32(32).int32(message.page);
    }
    if (message.taskId !== "") {
      writer.uint32(42).string(message.taskId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetBrokerReportResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetBrokerReportResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.brokerReport.push(BrokerReport.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.itemsCount = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.pagesCount = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.page = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.taskId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetBrokerReportResponse {
    return {
      brokerReport: globalThis.Array.isArray(object?.brokerReport)
        ? object.brokerReport.map((e: any) => BrokerReport.fromJSON(e))
        : globalThis.Array.isArray(object?.broker_report)
        ? object.broker_report.map((e: any) => BrokerReport.fromJSON(e))
        : [],
      itemsCount: isSet(object.itemsCount) ? globalThis.Number(object.itemsCount) : 0,
      pagesCount: isSet(object.pagesCount) ? globalThis.Number(object.pagesCount) : 0,
      page: isSet(object.page) ? globalThis.Number(object.page) : 0,
      taskId: isSet(object.taskId)
        ? globalThis.String(object.taskId)
        : isSet(object.task_id)
        ? globalThis.String(object.task_id)
        : "",
    };
  },

  toJSON(message: GetBrokerReportResponse): unknown {
    const obj: any = {};
    if (message.brokerReport?.length) {
      obj.brokerReport = message.brokerReport.map((e) => BrokerReport.toJSON(e));
    }
    if (message.itemsCount !== 0) {
      obj.itemsCount = Math.round(message.itemsCount);
    }
    if (message.pagesCount !== 0) {
      obj.pagesCount = Math.round(message.pagesCount);
    }
    if (message.page !== 0) {
      obj.page = Math.round(message.page);
    }
    if (message.taskId !== "") {
      obj.taskId = message.taskId;
    }
    return obj;
  },

  create(base?: DeepPartial<GetBrokerReportResponse>): GetBrokerReportResponse {
    return GetBrokerReportResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetBrokerReportResponse>): GetBrokerReportResponse {
    const message = createBaseGetBrokerReportResponse();
    message.brokerReport = object.brokerReport?.map((e) => BrokerReport.fromPartial(e)) || [];
    message.itemsCount = object.itemsCount ?? 0;
    message.pagesCount = object.pagesCount ?? 0;
    message.page = object.page ?? 0;
    message.taskId = object.taskId ?? "";
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}